Revise the pull request template
- Turns out, a checkbox list isn't the right format for that PR objective
section. It's more annoying to edit by hand, so we should reserve it for objects
where you'll want to update the state of the checkboxes after the fact.

- I've been getting annoyed at how much comment text I have to delete whenever I
open a PR, so I'm revising that blurb to fit completely within the default PR
description textarea height (200px) at a reasonable browser width. I think I've
kept all the same information without making it rude or anything.